Add `.env` support for docker-compose
Docker-compose supports setting variables using a `.env` file, so
instead of having to modify the original YAML file, this can be used
instead.

A sample `.env.example` is provided so the user can use it as a guide.
The `.env` file is ignored, as it's content is only relevant to the
user.