Got some "feedback" from someone over at plex
Remove the ports section, correct that host network mode negates this
Added three env vars for config, transcode, and data.  All default to ./volume/{} so that if one wanted to, they could get up and running quickly with:
```
mkdir -p ./volume/{config,data,transcode}
```
As far as the PLEX_CLAIM, it IS a secret.  The fact that it is ephemeral doesn't negate this fact.  It should be passed in on the command line and never spoken of again or in a secrets file.