Change chown to modify ownership symlinks
It is possible to end up with a long startup period if symlinks have incorrect ownership as they've never corrected by this script currently so each new docker pull they are found by find and the target ownership is updated (despite being correct) causing a delay in startup

This change assumes that all symlink targets are also within the /config directory tree otherwise the targets could be left unchanged as chown no longer follows the links. Alternatively we could run the command twice once for targets & once for symlinks themselves.