feat: add support for extra init and sidecar containers (#120)
The current chart asserts that only a single init container is needed for rclone. This can be adapted for most purposes so long as the specified image contains an entrypoint at /init/init.sh and is meant to be run as an initContainer and not a sidecar container (https://kubernetes.io/docs/concepts/workloads/pods/init-containers/#api-for-sidecar-containers).

There are useful tools that operate on plex's metadata files, bundles, or database that would be best run as sidecar containers (see above), which are functionally init containers with restart policies.

This pull request adds support for additional init containers through an "extraInitContainers" argument that is semantically identical to the existing extraContainers argument.